+++ lince/src/lince/core/app.h
@@ -111,17 +111,15 @@ float LinceGetAspectRatio();
 void LinceGetScreenSize(vec2 size);
 
 /** @brief Transforms a point in space from screen to world coodinates.
-* @param screen_coords vec2 with the x and y position on screen, overwritten with the result.
-* @param screen_size vec2 with the screen's width and height in pixels.
-* @param vp_inv Inverse of the view-projection matrix (obtained from a camera).
-* @todo Add extra argument for return vec2 world coordinates,
-*       instead of overwriting `screen_coords`.
+* @param world_pos vec2 that returns the XY position in the world.
+* @param screen_pos vec2 with the XY position in the screen to trasnform.
+* @param camera Camera with an up-to-date inverse view-projection matrix
 */
-void LinceTransformToWorld(vec2 screen_coords, vec2 screen_size, LinceCamera* camera);
+void LinceTransformToWorld(vec2 world_pos, vec2 screen_pos, LinceCamera* camera);
 
 /** @brief Transforms a point in space from world to screen coodinates.
-* @param screen_pos vec2 that returns the x and y position on screen.
-* @param world_pos vec2 with the x and y position in the world to transform.
+* @param screen_pos vec2 that returns the XY position on screen.
+* @param world_pos vec2 with the XY position in the world to transform.
 * @param camera Camera with an up-to-date view-projection matrix
 */
 void LinceTransformToScreen(vec2 screen_pos, vec2 world_pos, LinceCamera* camera);